refactor : Home의 Carousel 이름 변경
- Carousel ->InfinityCarousel
