+++ frontend/src/pages/HomePage/components/InfinityCarousel/styles.ts
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-export const CarouselContainer = styled.div`
+export const InfinityCarouselContainer = styled.div`
   position: relative;
 
   overflow: hidden;
