+++ frontend/src/pages/HomePage/components/ReviewMeOverview/index.tsx
@@ -3,7 +3,7 @@ import UsageCarosel1Icon from '@/assets/usageCarosel1.svg';
 import UsageCarosel2Icon from '@/assets/usageCarosel2.svg';
 import UsageCarosel3Icon from '@/assets/usageCarosel3.svg';
 
-import Carousel, { Slide } from '../Carousel';
+import InfinityCarousel, { Slide } from '../InfinityCarousel';
 
 import * as S from './styles';
 
@@ -32,7 +32,7 @@ const ReviewMeOverview = () => {
           <img src={WritingIcon} alt={OVERVIEW_TITLE} />
           <S.OverviewTitle>{OVERVIEW_TITLE}</S.OverviewTitle>
         </S.OverviewTitleContainer>
-        <Carousel slideList={OVERVIEW_SLIDES_LIST} />
+        <InfinityCarousel slideList={OVERVIEW_SLIDES_LIST} />
       </S.ColumnSectionContainer>
     </S.ReviewMeOverview>
   );
